Add project showcase with routing and details pages
Introduced react-router-dom for client-side routing. Added MyProjects and ProjectDetails components to display a list of projects and detailed project views. Created a MyProjectsData module for project data. Updated App.jsx to use routes and render the new components. Added related images to the public directory and wrapped the app in BrowserRouter in main.jsx.